Updates UI to use advanced splunk query features
Includes new datatypes for collecting data from bolt, splunk triggered actions,
puppet metrics, puppet facts.

Default landing page is now Inventory instead of Overview

Report Builder is added to the Reports menu, allowing users to browse reports filtered
by fact and value + report status, regardless of if they are building detailed reports
it also allows them to view the resulting query so they can use it when creating an alert
action.

A search and alert action is installed, but disabled, that will create a detailed report
for any report that indicates an atypical puppet run.

Beginnings of support for viewing PE metrics are present with the sourcetype

Includes a first draft of a README updates with more information

Updates application icon to be consistent