pie_67_macros Use macros to allow custom index
Before this update all the searches were configured assuming the main
index was the only option.

With this update, the searches will respect user set indexes for each
sourcetype or a single global puppet data index. See the advanced
configuration option in the Readme for more details on configuring it.

This also starts the 1.5.1 release process.