Rewrites alert actions to use a shared PIE library
Introduces puppet_bolt_action and puppet_report_generation functions which now do the heavy lifting of the actual alert handling (running the orchestrator, parsing data, etc). The modalert files are now just to generate python dictionaries that are passed to either of those two functions.

This changes authentication also, so instead of providing an auth token, a user provides username/password for corressponding access accounts, and we do the token generation for the user. This removes the headache of token expiration.