Pie 37 add bolt (#19)
* Merge task alert actions into report viewer

These are none working alerts at this moment because this file changes just updates the app to now be able to have multiple alerts to exist in the app and for the setup screens to collect the data needed for the alerts to function

* In progress save

Sketches out how to redo report generation with the common library

* Rewrites alert actions to use a shared PIE library

Introduces puppet_bolt_action and puppet_report_generation functions which now do the heavy lifting of the actual alert handling (running the orchestrator, parsing data, etc). The modalert files are now just to generate python dictionaries that are passed to either of those two functions.

This changes authentication also, so instead of providing an auth token, a user provides username/password for corressponding access accounts, and we do the token generation for the user. This removes the headache of token expiration.

* Adds working investigate function to PIE

debugalert.json is added to include information for triggering the investigate alert

This posts the alert back to to splunk, and includes the content of the original alert, however we may just want to link to it in future?

* Updates bolt action messages

Now contains status of the task

* Verified alerts work in Splunk

Both puppet_report_generation and puppet_bolt_action work when invoked by their corresponding modalert_*

* Tested changes. Some fixes were required.

* Removed unrequired import.

* Removed unrequired debug statement

* Removes modifications to the autogenerated files.

While not optimal, these files are autogenerated and shouldn't be modified except by the addon builder

* Updates readme to which files we modify

And which files we have to leave alone for the addon builder